Implement Tarantool iproto hello world
This patch implements basic iproto connection test. The app connects to
local Tarantool instance and pings it using Tarantool-c [^1] library
through FFI and `ocaml-ctypes` [^2] bindings.

Run Tarantool and configure it as follows.
```lua
box.cfg{listen = 'localhost:3301'}
```

Then run the app. The ping should be successful.

Note, `libffi` and `libtarantool` [^1] should be installed for the app
to be built.

[^1] https://github.com/tarantool/tarantool-c
[^2] https://github.com/yallop/ocaml-ctypes